REFACTOR : UI interface 분리
UI를 생성 시점에 변경하도록 UI와 구현하는 Console UI로 분리
