+++ docs/README.md
@@ -24,7 +24,7 @@
 <br>
 
 ### 리팩토링
-- [ ] 
+- [x] UI Interface 분리
 
 
 <br>
@@ -75,3 +75,7 @@
 
 * 구현하면서 역할에 따라 분리해서 진행하려고 했으나, 생각보다 너무 복잡해지고, 분리하는 것에 대해 명확한 기준이 생기지 않아서 먼저 동작하는 코드를 만들고자 했습니다.
 * 프로토타입 제작 이후, 사용자 입력시 발생할 수 있는 예외를 일괄 처리했습니다.
+
+**2023-10-23**
+
+* UI의 경우 지금은 Console 입력을 받지만 나중에는 사용자 입력을 받는 경우도 있을 것이라고 보고 UI를 Interface로 분리하고, 생성하는 시점에 정할 수 있도록 만들었습니다.
