+++ src/main/java/baseball/ui/UI.java
@@ -1,0 +1,17 @@
+package baseball.ui;
+
+public interface UI {
+    void startGame();
+
+    String insertNumber();
+
+    void gameResult(int[] ballAndStrike);
+
+    void winGame();
+
+    String selectGameProgress();
+
+
+
+
+}
